requirements.txt
Django==5.0.7
djangorestframework==3.15.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
requests==2.32.3
djoser==2.2.2